+++ src/utilities/get-next-id.js
@@ -1,0 +1,7 @@
+var counter = 0;
+
+export default function getNextId() {
+    counter += 1;
+
+    return counter;
+}