+++ web/app/store/application-notifications/application-notifications.reducer.js
@@ -1,0 +1,41 @@
+import immutable from 'immutable';
+import * as constants from './application-notifications.constants';
+
+let baseData = {
+  notifications: []
+};
+
+export default function(state = immutable.fromJS(baseData), action) {
+  let newState;
+
+  switch (action.type) {
+    case constants.ADD:
+      newState = state.setIn(['notifications'], state.getIn(['notifications']).push(immutable.fromJS(action.newNotification)));
+      break;
+
+    case constants.UPDATE:
+      newState = state.setIn(['notifications'], state.getIn(['notifications']).map((notification) => {
+        if (notification.get('id') === action.id) {
+          return notification.merge(action.updateData);
+        } else {
+          return notification;
+        }
+      }));
+      break;
+
+    case constants.REMOVE:
+      newState = state.setIn(['notifications'], state.getIn(['notifications']).filter((notification) => {
+        return notification.get('id') !== action.id;
+      }));
+      break;
+
+    case constants.CLEAR:
+      newState = state.setIn(['notifications'], state.getIn(['notifications']).clear());
+      break;
+
+    default:
+      newState = state;
+  }
+
+  return newState;
+};