+++ webpack.config.developer.js
@@ -61,6 +61,7 @@ module.exports = {
       'moment',
       'moment-timezone',
       'redux',
+      'redux-thunk',
       'reselect',
       'store-cacheable',
       'superagent',