+++ CHANGELOG.md
@@ -4,7 +4,12 @@ There are some key used within the changelog which are as following:
 
 - [BREAKING]: means there is a change that is known to break existing code
 
-## IN PROGRESS
+# IN PROGRESS
+- added utilities to managing auto incrementing id
+- added support to bind click event for the negative and position notification actions
+- added basic notification container with very basic example (example needs more flushing out)
+
+## 0.9.0
 
 - added extend text components with support for auto complete and tagging #116
 - added popover function included support for styling in the form of tooltip, popover, and drop down menu #131